requirements.txt
steam[client]
requests
openpyxl
tqdm
six
vdf
gevent
cachetools
pycryptodomex
protobuf==3.20.1
eventemitter
pycryptodome
oss2